+++ src/components/TeacherViews/SpecialistLessonDownloads/SpecialistLessonDownloads.fixture.ts
@@ -1,6 +1,5 @@
-import { SpecialistLessonDownloadsProps } from "./SpecialistLessonDownloads.view";
-
 import { LessonDownloadsData } from "@/node-lib/curriculum-api";
+import { SpecialistLessonDownloads } from "@/node-lib/curriculum-api-2023/queries/specialistLessonDownload/specialistLessonDownload.schema";
 
 export const downloads: LessonDownloadsData["downloads"] = [
   {
@@ -39,8 +38,8 @@ export const nextLessons = [
 ];
 
 const specialistLessonDownloadsFixtures = (
-  partial?: Partial<SpecialistLessonDownloadsProps>,
-): SpecialistLessonDownloadsProps => {
+  partial?: Partial<SpecialistLessonDownloads>,
+): SpecialistLessonDownloads => {
   return {
     lesson: {
       expired: false,